rename nerdss_analysis to analysis
